Introduce EventStore module
