+++ lib/eventory.rb
@@ -1,6 +1,5 @@
 require 'sequel'
 require 'eventory/version'
-require 'eventory/event_store'
 require 'eventory/event_data'
 require 'eventory/event'
 require 'eventory/recorded_event'
@@ -9,11 +8,11 @@ require 'eventory/persistent_subscription'
 require 'eventory/checkpoints'
 require 'eventory/checkpoint'
 require 'eventory/event_handlers'
-require 'eventory/event_builder'
 require 'eventory/event_stream_processor'
 require 'eventory/projector'
 require 'eventory/reactor'
 require 'eventory/domain'
+require 'eventory/event_store'
 
 module Eventory
 end
