--- lib/eventory/event_builder.rb
@@ -1,20 +1,0 @@
-module Eventory
-  class EventBuilder
-    def build(type:, data:)
-      klass = resolve_type(type)
-      if klass
-        klass.new(data)
-      else
-        data
-      end
-    end
-
-    private
-
-    def resolve_type(type)
-      Object.const_get(type)
-    rescue NameError
-      nil
-    end
-  end
-end
