+++ lib/eventory/event_store/event_builder.rb
@@ -1,0 +1,22 @@
+module Eventory
+  module EventStore
+    class EventBuilder
+      def build(type:, data:)
+        klass = resolve_type(type)
+        if klass
+          klass.new(data)
+        else
+          data
+        end
+      end
+
+      private
+
+      def resolve_type(type)
+        Object.const_get(type)
+      rescue NameError
+        nil
+      end
+    end
+  end
+end
