+++ spec/eventory/projector_spec.rb
@@ -20,7 +20,7 @@ RSpec.describe Eventory::Projector do
 
 
   subject(:test_projector) { test_projector_class.new(event_store: event_store, checkpoints: checkpoints) }
-  let(:event_store) { Eventory::EventStore.new(database: database) }
+  let(:event_store) { Eventory::EventStore::Postgres::EventStore.new(database: database) }
   let(:checkpoints) { Eventory::Checkpoints.new(database: database) }
   let(:namespace) { 'ns' }
